+++ resources/views/components/account-dropdown.blade.php
@@ -1,24 +1,17 @@
 @props(['btnClass' => null])
 
 <div class="dd cursor-pointer {{ $btnClass }}">
-
     @auth
-
         <div class="flex va-c py-025">
-
             @if (method_exists(auth()->user(), 'avatarUrl'))
                 <img class="wh-2 rounded-full" src="{{ auth()->user()->avatarUrl() }}" alt="Profile Photo">
             @endif
-
-            <div class="ml-075">
+            <div class="inline-flex va-c ml-075">
                 <span>{{ Auth::user()->name }}</span>
                 <x-gt-icon name="chevron-down" class="wh-1 ml-025" />
             </div>
-
         </div>
-
         <div {{ $attributes->class(['dd-content bx pxy-0 mt-025 block']) }}>
-
             @if (isset($content))
                 {{ $content }}
             @else
@@ -32,9 +25,6 @@
                     </x-gt-menu>
                 @endif
             @endif
-
         </div>
-
     @endauth
-
 </div>